feat(at): Add at and now. Move empty, never to own modules
