+++ packages/disposable/package.json
@@ -45,9 +45,9 @@
     "buba": "^4.0.2",
     "cpy-cli": "^1.0.1",
     "eslint": "^3.15.0",
-    "flow-bin": "^0.43.1",
-    "mocha": "^3.2.0",
-    "rollup": "^0.41.4",
+    "flow-bin": "^0.47.0",
+    "mocha": "^3.4.2",
+    "rollup": "^0.43.0",
     "rollup-plugin-buble": "^0.15.0",
     "rollup-plugin-node-resolve": "^2.0.0",
     "sinon": "^1.17.7",
