+++ packages/prelude/package.json
@@ -38,7 +38,7 @@
     "flow-bin": "^0.47.0",
     "mocha": "^3.4.2",
     "nyc": "^10.3.2",
-    "rollup": "^0.41.6",
+    "rollup": "^0.43.0",
     "rollup-plugin-buble": "^0.15.0"
   }
 }
