+++ packages/types/package.json
@@ -31,6 +31,6 @@
   "author": "brian@hovercraftstudios.com",
   "license": "MIT",
   "devDependencies": {
-    "flow-bin": "^0.43.1"
+    "flow-bin": "^0.47.0"
   }
 }
